feat: move diff html function to separate file
