+++ scripts/checkArticles/index.js
@@ -11,7 +11,6 @@ const chalk = require('chalk');
 const queryString = require('query-string');
 const fetch = require('isomorphic-fetch');
 const fs = require('fs');
-const Differ = require('diff-match-patch');
 const mkdirp = require('mkdirp');
 const { getNdlaApiUrl } = require('../../src/config');
 const {
@@ -19,9 +18,9 @@ const {
   learningResourceContentToHTML,
 } = require('../../src/util/articleContentConverter');
 const { resolveJsonOrRejectWithError } = require('../../src/util/apiHelpers');
+const diffHTML = require('./diffHTML');
 
 const dom = new jsdom.JSDOM('<!DOCTYPE html></html>');
-const differ = new Differ();
 
 global.window = dom.window;
 global.document = window.document;
@@ -130,75 +129,6 @@ async function fetchAllArticles(url) {
     articleIds.push(...articles);
   });
   return articleIds;
-}
-
-/**
- * Get current, next and previous diff values. Return undefined if one of them is undefiend
- */
-function getValues(index, diffs) {
-  const prevDiff = diffs[index - 1];
-  const nextDiff = diffs[index + 1];
-  const diff = diffs[index];
-
-  if (prevDiff === undefined || nextDiff === undefined || diff === undefined) {
-    return undefined;
-  }
-  return {
-    current: diff[1],
-    next: nextDiff[1],
-    previous: prevDiff[1],
-  };
-}
-
-// I.E "<h2>Oppgaver</h2> <ol>...</ol>" -> "<h2>Oppgaver</h2><ol>...</ol>"
-function allowSpaceRemovalBetweenTags({ current, next, previous }) {
-  return (
-    previous[previous.length] !== '>' && current === ' ' && next[0] === '<'
-  );
-}
-
-// I.E "<table><tbody>...</tbody></table>" -> "<table><thead>...</thead><tbody>...</tbody></table>"
-function allowTHeadInsertion({ current, next, previous }) {
-  return (
-    previous.endsWith('<table><t') && current === 'body' && next === 'head'
-  );
-}
-
-function isRemovalAllowed(index, diffs) {
-  const values = getValues(index, diffs);
-  if (values) {
-    const result = [allowSpaceRemovalBetweenTags, allowTHeadInsertion].find(
-      fn => fn(values) === true,
-    );
-    return result !== undefined;
-  }
-  return false;
-}
-
-function diffHTML(oldHtml, newHtml) {
-  const diffs = differ.diff_main(oldHtml, newHtml);
-  differ.diff_cleanupSemantic(diffs);
-  let diffString = '';
-
-  let shouldWarn = false;
-
-  diffs.forEach((diff, index) => {
-    // green for additions, red for deletions
-    // grey for common parts
-    const [result, value] = diff;
-    if (result === 1) {
-      diffString += `${chalk.green(value)}`;
-    } else if (result === -1) {
-      diffString += `${chalk.red(value)}`;
-      // Some diffs are allowed
-      if (!isRemovalAllowed(index, diffs)) {
-        shouldWarn = true;
-      }
-    } else {
-      diffString += value;
-    }
-  });
-  return { diff: diffString, warn: shouldWarn };
 }
 
 function testArticle(article) {
